Feature/dirac mpi fix (#30)
The dirac unit test is modified to work on multiple processors.

The test now takes latitude/longitude pairs as inputs for the forcing locations. Using a nearest-neighbor search, the dirac initilization for MPAS determines the cell and processor to which the user-selected locations belong. The JSON file is updated to include the original cell locations from earlier tests (cells 100 and 2 of the 2562 cell MPAS mesh). Additional latitude and longitude pairs are given for the 20 largest world cities for customizable testing with larger numbers of forcing locations.

The increment norms for Ldelta and Bdelta are similar to those from the previous version, and are not updated here. It should be noted that having nl0r set to 3 with the "cor" method for localization increases the memory requirement substantially, and even more so when 2 processors (>12 GB) are used instead of a single processor (>4 GB). There still remain small differences between single and 2 processor outputs for L*delta at the grid-scale.


 Changes to be committed:
	modified:   model/mpas_fields_mod.F90
	modified:   model/mpas_geom_mod.F90
	modified:   test/testinput/dirac.bump.json