Feature/radiance updates (#46)
Radiance and AOD observations now require a channel list in the config file. The UFO radiance operator also requires specific instrument names in the config file, which allows multiple instruments to be used.

This fixes the hofx test for mpas and corresponds to UFO PR's JCSDA/ufo#114 and JCSDA/ufo#127.