Remove legacy dependency on Boost unit test framework (#106)
This change removes the mpas dependency on the Boost unit test framework. The ctests no longer use this Boost framework. This change is just cleanup right now but will be required soon in Singularity when Mark's new containers are released, as they will no longer contain the Boost test framework. This change has been tested in both the current Singularity image and the upcoming one.