Add hyperspectral IR radiances to hofx and hofx3d (#107)
AIRS, CRIS, and IASI are added to the hofx and hofx3d ctests. Currently using the same channel selection as FV3.  We are using USGS land type categories in our MPAS simulation, which requires specifying USGS for IRVISlandCoeff.  Also added "ghost" entries for ATMS that can be uncommented once ATMS ctest is turned on for UFO.