Put valid time on x-axis for SNAPSHOTFC figures (#223)
TYPE: enhancement

KEYWORDS: lead time, graphics, SNAPSHOTFC

DESCRIPTION OF CHANGES: 
Previously the analysis time was used on the x-axis of the SNAPSHOTFC figures.  It is more intuitive to put the valid time on the figure.  Part of the solution is to allow each line in plotTimeSeries to have a unique set of x-values.  The old behavior of common x-values between all lines is also maintained.

The x-axis dates of all SNAPSHOT figures are modified to use concise formatting.  The figures will be more readable.

ISSUE:
None

LIST OF MODIFIED FILES: 
M       test/graphics/plot_stats_timeseries.py

TESTS CONDUCTED: Script was run with refactored code and produces expected results.