Remove gen_ens_pert_B test (#231)
TYPE: bug fix

KEYWORDS: ctest

DESCRIPTION OF CHANGES: Currently the gen_ens_pert_B ctest produces different output when performed with a Debug build and a Release build. This means the 'Compare' part of the test will always fail with at least one of the builds since the reference output needs to be set for one or the other. This change removes the test to prevent it from generating failures that must be ignored. It can be restored later if and when it can produce the same output for both build types.

LIST OF MODIFIED FILES:
M test/CMakeLists.txt

TESTS CONDUCTED: Tested in Singularity