remove obsolete config "no_transf" in read routine (#233)
TYPE: maintenance

KEYWORDS: cleanup

DESCRIPTION OF CHANGES: Removing the obsolete configuration "no_transf" in read routine. For a while ago, a variable `temperature` was **directly read** from file for `ensemble` (which was a format of restart), while it was **diagnosed** after being read from file for `background`. Now we process it consistently both for background and for ensemble, so this `no_transf` is obsolete and removed. This PR does not make any change in results.

LIST OF MODIFIED FILES: 
M       model/mpas_field_utils_mod.F90

TESTS CONDUCTED: Tested in Singularity.