Encapsulate the statistics in a database container (#260)
TYPE: enhancement

KEYWORDS: refactoring, database, statistics

DESCRIPTION OF CHANGES: 
This is the next PR in a series of major refactorings of the observation space diagnostic plotting.  As before, the statistics across multiple cycle times, forecast lengths, experiments, observation spaces, variable names, and binning parameters are stored in a pandas DataFrame.  This PR places the initialization of that DataFrame in the constructor of a new class called `StatsDataBase`.  
Additionally, a new class called `DFWrapper` is created to wrap a DataFrame for our purposes, and specifically to make slicing the DataFrame easier in an 8-dimensional space.  This is achieved by enabling the DataFrame to be sliced using a dictionary instead of a tuple.  A tuple is generated from the Dictionary to be given as an argument to the DataFrame.loc method and includes `slice(None)` for non-specified axis members of the `DFWrapper` index.

The versatile and non-ordered functionality of the dictionary slicing method makes it easy to re-order loops and create new plot types without needing to worry about the order of axes selected in a tuple.

ISSUE: 
None

LIST OF MODIFIED FILES: 
A       test/graphics/StatsDB.py
M       test/graphics/binning_utils.py
M       test/graphics/config.py
M       test/graphics/plot_stats_timeseries.py
M       test/graphics/plot_utils.py

TESTS CONDUCTED: all figures were re-generated for 5 different ObsSpaces.