Add EDA application and utilize Geometry-specific MPI communicator (#389)
TYPE: new feature

KEYWORDS: eda, geometry, communicator

DESCRIPTION OF CHANGES:
This PR adds an initial EDA implementation and an application ctest.  Part of enabling EDA is to allow for each Geometry object to have an independent MPI communicator.  The communicator is passed down through mpas_init so that each Geometry object can carry out independent communications without overwriting memory occupied by other Geometry objects.

Part of the Geometry refactoring is to change the copy constructor so that the LHS refers to "self" and the RHS refers to "other" in the Fortran code.  That makes the nature of the copy constructor less confusing.

For the new ctest (`eda_3dhybrid`), the reference output is added as a link to `3dhybrid_bumpcov_bumploc.ref`, which has an identical output on the head processor.  So far no reference or validation is added for other eda members.

For EDA, there is a change to how the config "member" entry is used to generate filenames and filepaths using the %{member}% placeholder in yaml files.  In a follow-on PR (#386), that member config entry might be removed with a requirement that the actual member string, not a placeholder, be explicitly included across multiple yamls.  Those changes are saved for the follow-on PR.

It should be noted that moving the mpas_init call to the Geometry constructor will result in a memory leak across multiple outer iterations.  That is because `mpas_finalize` is not compatible with mpas-jedi use-cases as currently written and thus is not called.  We need to figure out a better way to deconstruct the MPAS-Model `corelist` and `domain` for mpas-jedi that does not finalize the MPI communicator (as described in #390).

ISSUE: 
Closes #317 
Closes #319 

LIST OF MODIFIED FILES:
M       mains/CMakeLists.txt
A       mains/mpasEDA.cc
M       model/Fortran.h
M       model/GeometryMPAS.cc
M       model/getvalues/mpasjedi_getvalues_mod.F90
M       model/mpas_field_utils_mod.F90
M       model/mpas_geom_interface.F90
M       model/mpas_geom_mod.F90
M       model/mpas_run_mod.F90
M       test/CMakeLists.txt
A       test/testinput/eda_3dhybrid.yaml
A       test/testoutput/eda_3dhybrid.ref

TESTS CONDUCTED: Ctests pass on cheyenne with latest gnu-openmpi module.  EDA application also has been used in multi-day cycling experiments.