Reduce fg-check threshold for sfcp in 3dfgat, update ref (#457)
TYPE: bug fix

KEYWORDS: 3dfgat, background check, pressure_p, hydrometeors, intel build

DESCRIPTION OF CHANGES:
Motivation - There are a large number of surface pressure observations removed by the Background Check in the second outer iteration of the 3dfgat ctest, possibly owing to the unrealistic B or other simplifications.  The number of filtered surface pressure observations is sensitive to the simulated h(x) values.  In this case with a time-varying MPAS forecast state, the h(x) values vary more widely across different machines and compilers, causing a different number of observations to be filtered and wildly different cost function values. The numerical results are described in further detail in #452 and #424.

Solution - The simplest way forward for now is to reduce the first-guess threshold.

Second change - index_qc and index_qi are removed from the analysis variables, because none of the assimilated observations in 3dfgat are sensitive to them. index_qr, index_qs, index_qg, pressure_p are added to model/state variables to ensure they are reset to the background values in the second outer iteration.  This change was found to have an impact on Radiosonde CostJo in the second iteration.

ISSUE:
Start to solving #424 

LIST OF MODIFIED FILES:
M       test/testinput/3dfgat.yaml
M       test/testoutput/3dfgat.ref

TESTS CONDUCTED: 3dfgat ctest with gnu and intel on Cheyenne.  Passes CI-Intel tests, while develop did not.