Add correct array boundaries (#476)
TYPE: no impact

KEYWORDS: lineargetvalues

DESCRIPTION OF CHANGES: While looking at another issue, we noticed a place where an array is passed to a function without strictly correct array bounds. We thought for a time that fixing this error caused a change in many ctest results, but that turned out not to be the case. However, it is still good practice to pass the correct array bounds.

LIST OF MODIFIED FILES:
M src/mpasjedi/getvalues/mpasjedi_lineargetvalues_mod.F90

TESTS CONDUCTED: ctests with gnu-openmpi on Cheyenne, and CodeBuild