Uncomment default issue message in PR template (#494)
TYPE: no impact

DESCRIPTION OF CHANGES:
This PR removes the KEYWORD section and the comment/quote encapsulation of the issue to be closed by a PR.  The KEYWORD section is not really meaningful for us.  Having the issue number commented prevents the issue from being closed when the PR is merged.  It is easier to keep this procedure automated.  If a PR does not fix/close an issue, then a different word besides `Fixes` or `Closes` can be used, such as `Addresses` or `Related to`.

ISSUE:
None