Ensure fixed order of diagnostics in statistics files between experiments (#539)
TYPE: enhancement, no impact

DESCRIPTION OF CHANGES:
When comparing statistics files using nccmp and ncdiff, the order of the rows needs to be identical.  This change ensures that the diagnostics `diagName` variable is in the same order between experiments by sorting the dictionary alphabetically in a single for loop.

There is no downstream impact on plotting, because the plotting scripts read in all variables row-by-row, then sorts them in a pandas DataFrame using the `diagName` value, and not the row ordering.

Here is an example of the diagName output using ncdump:
```
 diagName = "oma", "oma", "oma", "oma", "oma", "oma", "oma", "oma", "oma", 
    "oma", "oma", "oma", "oma", "oma", "oma", "oma", "oma", "oma", "oma", 
    "oma", "oma", "oma", "omb", "omb", "omb", "omb", "omb", "omb", "omb", 
    "omb", "omb", "omb", "omb", "omb", "omb", "omb", "omb", "omb", "omb", 
    "omb", "omb", "omb", "omb", "omb", "rltv_oma", "rltv_oma", "rltv_oma", 
    "rltv_oma", "rltv_oma", "rltv_oma", "rltv_oma", "rltv_oma", "rltv_oma", 
    "rltv_oma", "rltv_oma", "rltv_oma", "rltv_omb", "rltv_omb", "rltv_omb", 
    "rltv_omb", "rltv_omb", "rltv_omb", "rltv_omb", "rltv_omb", "rltv_omb", 
    "rltv_omb", "rltv_omb", "rltv_omb", "sigmaoa", "sigmaoa", "sigmaoa", 
    "sigmaoa", "sigmaoa", "sigmaoa", "sigmaoa", "sigmaoa", "sigmaoa", 
    "sigmaoa", "sigmaoa", "sigmaoa", "sigmaob", "sigmaob", "sigmaob", 
    "sigmaob", "sigmaob", "sigmaob", "sigmaob", "sigmaob", "sigmaob", 
    "sigmaob", "sigmaob", "sigmaob" ;
```
Here the order is `oma`, `omb`, `rltv_oma`, `rltv_omb`, `sigmaoa`, `sigmaob`, which is alphabetical.  Previously there was no guarantee that the diagnostics were sorted in any specific order.

ISSUE:
None

LIST OF MODIFIED FILES:
M       graphics/DiagnoseObsStatistics.py

TESTS CONDUCTED:
Tested for verification in a cycling experiment.  In cases other than the one stated, there is no impact.