Do not post-process variables with all missing data (affects ABI/AHI) (#584)
TYPE: bug fix

DESCRIPTION OF CHANGES:
The JediDB.FileHandles.varsInGroup method is used to determine all of the variable names belonging to an ObsGroup, including separating radiance data by channel number.  When a channel is not assimilated, the new IODA-v2 code writes out all missing values for those channels.  So in post-processing we do not need to handle those missing data.  This PR ensures that only variables with some finite values are processed.  Without this PR, DiagnoseObsStatistics.py no longer works for ABI and AHI when we select a subset of channels (typically only in variational runs, not in hofx).

ISSUE:
None

LIST OF MODIFIED FILES:
M       graphics/JediDB.py

TESTS CONDUCTED: Tested that the modification works for ABI and AHI in a single variational cycle when assimilating channels 8-10 and the input/output observation files contain channels 7-15.  Note, this will also affect post-processing for other observation types when all-missing data is present in the feedback file for any one variable.