Uncommented the "obs space.obsdataout" specs in the 4DEnVar test configuration YAML files. (#589)
TYPE: bug fix

DESCRIPTION OF CHANGES:

The changes in this PR are simply to uncomment the obdataout specs in the two 4DEnVar tests.

This PR enables the writing of the ioda output files (ie, uncomments the obsdataout specs) in the two 4DEnVar tests that were exhibiting crashes when attempting to write the ioda output files. There is a companion PR in ioda (jcsda-internal/ioda/pull/367) that has the fix for this crash.

Note that this PR needs to be merged with or after the companion ioda PR (jcsda-internal/ioda/pull/367).

ISSUE:
Fixes #575


LIST OF MODIFIED FILES:

M       test/testinput/4denvar_ID.yaml
M       test/testinput/4denvar_bumploc.yaml

TESTS CONDUCTED:

Ran the mpas-jedi 4denvar tests on my iMac system and these tests failed (write crashed) without the ioda fix, and passed with the ioda fix (including the successful writing of the ioda output files).