Turn on 2D binning for all-sky and land fraction for AHI/ABI (#909)
TYPE: maintenance

DESCRIPTION OF CHANGES: This PR makes default the 2D binning for all-sky and land fraction categories for AHI/ABI. This is done by turning `True` the variables `specialAllSkyBins` and `specialLandFracBins` in `config.py` created in #858. However, as a requirement, ydiags and geovals need to be written for each DiagSpace. At this moment, this is only for AHI/ABI but it can be extended for other radiance data given that geovals are written during the DA. 

LIST OF MODIFIED FILES:
M       graphics/config.py
M       graphics/var_utils.py

TESTS CONDUCTED: 
- Tested on Cheyenne using the MPAS-Workflow