Changes for ewok and enabled correlated obs error (#880)
TYPE: no impact

DESCRIPTION OF CHANGES: A few changes to make the mpas interface work with ewok. One is adding the factory so we can use correlated obs error, the other is to use a configuration instead of parameters for State::write because the Parameters give an unknown exception which we were not able to figure out.