Add obs error factory instantiation to hofx (#1078)
TYPE: bug fix

DESCRIPTION OF CHANGES: The obs error factory was not instantiated in hofx (but was in hofx 3d). This means the obs error could only be diagonal, and users couldn't run with `cross variable covariances` for example. 
This just adds the instantiation of the factory in the main. 

LIST OF MODIFIED FILES:
M       src/mains/mpasHofX.cc

TESTS CONDUCTED: ctests in mpas-jedi, hofx with `cross variable covariances` in skylab