Fix saber block configuration in spectral test (#1097)
TYPE: bug fix, no impact

DESCRIPTION OF CHANGES: The setup of the interpolation block in the original test had the localization matrix and the ensemble members on different geometries. This moves the application of the interpolation to the correct location so the localization and ensemble members are now on the same grid.

LIST OF MODIFIED FILES:
M       test/testinput/dirac_spectral_1.yaml
M       test/testoutput/dirac_spectral_1.ref

TESTS CONDUCTED: ran ctests